- add card service tests - add move card endpoint

+++ backend/src/main/java/com/backend/card/api/UpdateCardRequest.java
@@ -2,8 +2,6 @@ package com.backend.card.api;
 
 import jakarta.validation.constraints.NotBlank;
 
-import java.math.BigInteger;
-
 public record UpdateCardRequest(
         @NotBlank
         String title,
@@ -11,6 +9,6 @@ public record UpdateCardRequest(
         String description,
 
         @NotBlank
-        BigInteger version
+        long version
 ) {
 }
